Update patch set 1
Patch Set 1:

That delay is nothing new. Start an Eclipse with EGit 5.6.0 or 5.5.1 and click through the repositories in the repo view. First time you also see that delay until the decoration appears. Also when you close and re-open the view; then you'll see that delay again.

Thereafter one didn't see it anymore because of that "adding a blank" hack. If we want to avoid it on later switches, we'll have to keep a map of ViewerLabels and unconditionally re-use the old label before calling updateTitle again.

Patch-set: 1
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0