Upgrade project properties to 2019-12 format
Open and save all project property pages to have new values stored
explicitly in version control. There are no functional changes (i.e. all
lines must be added, or must be a version increment).

This change does not guarantee that the properties are the same for all
projects in the repository.

Change-Id: I39e0bf55811d27ba71ae2adda9010ac26b60bc75
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>