Execute cleanup "Push down negations"
Push down negations to make reading of complex expressions with
negations easier.

Additionally to running the cleanup I intended to enable the
corresponding setting in the project properties, but unfortunately there
is a bug in the cleanup, which would lead to non-compiling code in egit,
if it were enabled: https://bugs.eclipse.org/bugs/show_bug.cgi?id=558635

Therefore we can only enable the cleanup in the properties after that
bug has been fixed and a new Eclipse version is released.

Change-Id: I1a5026304b080b0ccebe06e6829857b784541f99
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>