Update patch set 1
Patch Set 1:

I'm updating the setup archiver so that it will produce more stable results.  Part of this involves checking that the loaded resource does not contain unrecognized XML content, so that called me attention to this setup.  These changes clean up the setup to remove XML attribute and elements that are ignored anyway.

Patch-set: 1