Update patch set 3
Patch Set 3:

The implementation should now be in state where it can be reviewed functionally as well as formally.
I am not sure about the "can bulk create" condition, yet. Currently it is required that there is no change in any of the selected repositories. This may be too strict as simple edits or new files do not prevent a successful branch creation for a single repository.

Patch-set: 3
Label: Verified=0