Add Orbit version used in target platform to EGit repository
This ensures that EGit can be installed with only the EGit update
site being activated in Eclipse. If any dependencies are missing
locally in the EGit repository, p2 will try to fetch them from the
linked repositories, even if the user has unchecked "Check all update
sites" in the installation dialog.

Bug: 559735
Change-Id: If2587a76045ca5c29992b2348076d0a6cd5bbd9b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>