Update patch set 2
Patch Set 2:

> I tried this version and found the installation doesn't set the
 > selected target platform so I had to set it manually. Not sure why
 > this happens.

I'm not exactly sure what you mean, but definitely if you try to perform in an existing workspace with this modified setup you'll be prompted for the EGit Target Platform, and can choose.  The one you've been using is the first choice so will be selected by default.  Once you've performed, and saved this variable as a result, it won't be prompted again in that workspace.

For an initial setup, the user/contributor will also be prompted, but then too, the value will be saved and used for his workspace setup.

So I expect all is good and fine in this regard.

Note that to test that the setup indeed doesn't import the current unwanted projects, you can delete all the projects from the workspace and perform the import task again to see that the projects with m2e nature are not imported...

Patch-set: 2
Reviewer: Gerrit User 1220 <1220@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0