Extract a stand-alone DiffEditor
Extract all the non-page stuff from DiffEditorPage into a stand-alone
top-level editor part that can be used to show externally generated
DiffDocuments.

Extend the DiffEditorInput to also record the commits. Change the
internal logic of DifFEditor such that it doesn't change its input;
this is important if one wants to find an already open DiffEditor.

Bug: 544510
Change-Id: I73d5c5404ca26c129e6948401959eb0debc7e348
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>