Avoid manifest marked modified during broken build
When the target platform is broken, then the build will remove the OSGI
service declarations from the manifest. Later with a correct target
platform they are re-added, but at the end of the manifest.

Check in this modified version to avoid the manifest being marked as
changed during future occurrences of such delete-add-cycles.

Change-Id: Ibfcd229c8695c6489dd2a42fd126fc3955f15e23
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>