[repo view] Border for inline renaming of groups
Inline renaming in the project explorer adjusts the size of the text
widget to the contents and draws a border around it. Make the inline
editor in the git repositories view match that visual appearance and
behavior.

Code inspired by org.eclipse.ui.actions.RenameResourceAction but
adapted to using the JFace CellEditor framework instead of plain SWT.

Change-Id: I63d08682846598a4ee06f980860c907068185f11
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>