Update patch set 5
Patch Set 5:

There's a problem with the GitRepositories.isBare property tester (RepositoriesViewPropertyTester). The plugin.xml uses

  <test property="GitRepository.isBare" value="false" />

but that matches unless _all_ repos in the group are bare. If there's one non-bare repo in the group, the <test> will succeed and the command will be called with bare repos. Looks like this tester now needs an additional (optional) parameter for the operator ("or" or "and"), and here it should be "or".

Patch-set: 5