Update patch set 5
Patch Set 5:

As far as I can see, all isBare checks in the plugin.xml are against value=false. In almost all cases there is an additional count test (almost always 1).
I agree that there is an issue. But shouldn't it be dealt with separately? In which cases (for SwitchToMenu) do we have a problem?

* Only if more than one repo is involved (otherwise the property tester works as expected).
* Create new Branch? No, because of canBulkCreateNewBranch!

This leaves switching the branch to one of the common branches, but with respect to this feature, nothing changed at all.

I could let createMostActiveBranches create nothing if one if the repositories is bare. In this case, we'd run into createDisabledMenu which is correct. And as a user, I would not complain about the SwitchToMenu still being shown!

Patch-set: 5