Update patch set 6
Patch Set 6:

(1 comment)

> The commits are now sorted by time for comparison.

Changed to use the older revision as base. Checked with comparing the v5.6.1 tag against this commit: CompareCommand must be shown as added.
Made sure that "Compare with Each Other" and "Compare in Tree" agree on the direction of the comparison. (CompareTreeView.setInput() seems to have its parameters named wrongly.)

There's still a mess with the editor titles/view descriptions; ideally they should say "Comparing <base> with <newer>", but that is inverted. To be fixed in a later commit.

Patch-set: 6