Update patch set 6
Patch Set 6:

I am not sure about the change in the property tester. This particular property is used only for the switch menu. The justification for claiming that a selected group counts as multiple repositories is that  a group generally represents multiple repositories. An or-branch in the plugin.xml would be possible as well.

... Why can't the SwitchMenu determine its own label? Basically, 30 lines of plugin.xml code are duplicated just to add "Repositories" to the label...

Patch-set: 6
Label: Verified=0