Update patch set 6
Patch Set 6:

(2 comments)

I'd leave the property tester unchanged. See plugin.xml, PS 5, for my suggestion for a simple fix.

 > ... Why can't the SwitchMenu determine its own label?

If you know a way by which a dynamic sub-menu can programmatically change the label of the parent's MenuItem: feel free to change this. We did it that way because we hadn't found a way to do so. There is IElementUpdater, but that's only for handlers, not for dynamic contributions?

Patch-set: 6
Label: Verified=0