Update patch set 7
Patch Set 7:

Patch 7 is based on patch 5. I added the proposed property tester to the single repository switch case. If a group with a single repository is selected, the menu for single repositories is shown (with other submenu elements). I hope, this does not cause confusion ("But I selected a group!?"). If the group is not expanded, the label still indicates that it contains only a single repository.

If the parent "switch to" item was a dynamic contribution itself, you would have control over the label, wouldn't you?

Patch-set: 7
Label: Verified=0