Update patch set 1
Patch Set 1:

This is a first attempt to make the Push To Upstream label dynamic. I left the orignial menu item in place for now (for comparing visibility and enablement).

Feedback as to the actual format of the label would be very welcome:
* remote name
* push-URIs (what if there are many?)
* pushRefSpecs (what if there are many?)

One problem is that the width of the repositories view context menu is not reduced again, even if the push item is not shown in the end (may be caused by getContributionItems being called and always returning items).

Patch-set: 1