Update patch set 1
Patch Set 1:

I'm confused by the new menu, since it now shows a URL, that I cannot recognize. In my mental model, "Push to upstream" is just a shortcut for the "Push branch..." wizard, with one of the available remotes pre-selected. That's why I would rather suggest the label of the remote there, e.g. "Push to origin" in many cases. Or am I wrong about what the functionality of the current "Push to upstream" is?

The menu size was already reported in https://bugs.eclipse.org/bugs/show_bug.cgi?id=540778 and also happens with long branch names. If the new menu were to show the remote name, then this is probably not much of an issue, since I imagine typical remote names to be short (much shorter than some branch names, which that bug was about).

Patch-set: 1