Update patch set 1
Patch Set 1:

The current state really is only about being able to affect the label. Is "Push to upstream" really a shortcut for "Push branch"? The operation that is called pushes all configured refMaps to all configured push-URIs (at least that is what I read from the code.
This is why I am not sure about the label. "Push to <remoteName>" is more explicit than "upstream", but is it enough?

If the menu width is not really an issue, I will update the change accordingly.

Patch-set: 1