Update patch set 1
Patch Set 1:

Thanks for the IElementUpdater hint. I adapted the code accordingly. As both handlers already used SimpleConfigurePushDialog for obtaining the remote config, I placed the method for computing the push-to label there.

For now, only the remote name is shown. The question whether to indicate when several push URIs and mappings are present, is still open.

Patch-set: 1
Label: Verified=0