Update patch set 4
Patch Set 4:

In Patch Set 4, I attempt to fix the tests and restore the simple label for push/remote node.

The fetch labels will be dealt with in a further patch set. I agree that the labels should be consistent, just wanted to make sure that this is to be done in the context of the ticket.

Patch-set: 4
Label: Verified=0