Use a job to run GitProjectData.reconfigureWindowCache();
Currently the EGit activator slows down the startup of the IDE, moving
GitProjectData.reconfigureWindowCache(); to a job seems to improve the
startup as reported by the OSGi startup logging.

To ensure things are initialized, we wait at the end of the start method
for the job to finish.

To check for yourself, in your runtime configuration select the Trace
tab, enable it, select org.eclipse.osgi and select debug and
debug/BundleStartTime

Bug 560412
Change-Id: I627d2fdf810e6f5b8458a47c7dcbbf641cc2d6ba
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>