Update patch set 1
Patch Set 1:

What about instead of using an E4 event that gets posted post-startup, we bind it to the IWorkspaceService or IWorkbench?

{org.eclipse.core.resources.IWorkspace}={service.id=194, service.bundleid=90, service.scope=singleton}
  "Registered by bundle:" org.eclipse.core.resources_3.13.600.v20191122-2104 [90]
{org.eclipse.e4.ui.workbench.IWorkbench}={service.id=228, service.bundleid=118, service.scope=singleton, id=5392e09d-575c-4937-a619-40e193c52ff8}
  "Registered by bundle:" org.eclipse.e4.ui.workbench_1.11.0.v20191120-1917 [118]
  "No bundles using service."

The idea would be to have a setWorkspace on the repository change csanner, and it could use that instead of PlatformUI.getWorkbench() instead. You'd then have the lifetime of the service based on the existence of the workbench, and it would be shut down when the workbench goes away (or implicitly when the containing bundle is stopped)

Patch-set: 1