Move unnecessary jobs to DS components
This move the AutoShareProjects setup to a DS component, which gets kicked off
by the existence of the IWorkspace and IJobManager services when available. As
a result, we don't block the main thread from starting up, and can dismiss the
splash screen quicker.

THIS IS A WORK IN PROGRESS - DO NOT MERGE

Bug: 560412
Change-Id: I40ac12b5b95157c1de34bf471d508cfdb305fe47
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>