Update patch set 1
Patch Set 1: Code-Review-1

This is a WIP, so shouldn't be merged as-is.

However, we can shave off another 30ms with this change, and I think there are more opportunities to move to other background jobs as well. We might be able to use @Component annotations but I'd need to check that this worked with the existing build infra and not just PDE before doing that.

I'll do some more work on this change and see if we can pull it out of the list.

Patch-set: 1
Reviewer: Gerrit User 3 <3@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1