[repo view] Paste Repository URI menu items
It is confusing that the "paste" menu contribution is shown everywhere
but the corresponding clone repository action (for which it is a short
cut) is not.
After this change the paste command is still always active, but only
shown in the context menu for an empty or single repository group
selection. It is also added to the view menu.

Bug: 561016
Change-Id: I45448536cbd094c82a17d3dc4c7f8f285cc32eca
Signed-off-by: Alexander Nittka <alex@nittka.de>