WIP investigate TagActionTest Failure
The purpose of this change is to investigate the TagActionTest failures
of change 138436. Currently, I cannot see the modified tag sort order
influencing the tests and locally (in Eclipse) all the tests pass.
By removing most of the other tests and adding asserts/logging, I want
to reduce build time and find out, what makes the tests fail.

PS 5: are we sure that the bot selected the correct table item?

Change-Id: I1b592e50add9bf835c3a3d4bca3a7ec44faea72b
Signed-off-by: Alexander Nittka <alex@nittka.de>