Update patch set 1
Patch Set 1:

> @Thomas: Could you give me a brief feedback

Personally I don't use that paste often. It's a very old feature I think. It's probably always enabled because when the view is set to follow the current selection in other editors or views, the selection very frequently will be a file in the working tree, not a repo or repo group. So sometimes the paste would work, and sometimes not. Having to figure out that this only works when a repo or repo group (or nothing at all) is selected is not obvious.

If the problem is the visibility of the command in the menu, its visibility could perhaps be restricted. But that would give us a situation where pasting would actually work but the command wasn't shown in the menu. Not sure that's any better.

If the problem is that a paste too often results in a "Cannot create git repo from clipboard content" dialog, then perhaps we could change the handling of this message. (Log only, or show in the status bar, or have a self-vanishing popup in the repo view, or something like that. Or perhaps just silently ignore.)

Patch-set: 1
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0