Update patch set 1
Patch Set 1:

Thanks for the feedback. I'd have no real problem leaving everything as it is. I just felt that there is an inconsistency in the context menu entries. For me the paste is a short-cut for "clone repository and have the URI field already entered, so why do I see one but not the other? Well, for historical reasons.

I wanted to argue that it would not hurt to change the visibility, but indeed the paste is not part of the view menu where all the add/clone/create repository commands are present independent of the selection.

Patch-set: 1
Reviewer: Gerrit User 236689 <236689@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0