Update patch set 7
Patch Set 7:

Rebase in PS 6 as well as making the sort direction field volatile. PS 7 adds an additional assert in one of the three tests that sometimes seem to fail. Locally, I was not able to reproduce the test failure.

Note that whereas in build 1335 HistoryViewTest failed, in build 1344 it did not.

Patch-set: 7
Label: Verified=0