Update patch set 10
Patch Set 10:

All the screenshots in https://ci.eclipse.org/egit/job/egit.gerrit/1360/#showFailuresLink show a tag we did not want to select. Even if the code supposedly clears any prior selection and text and selects the tree item only if the label fits. Locally, the tests pass (Eclipse, Windows 10). The failing asserts seem to indicate that the screenshots are actually made when the test failed... But where are the other tags? Maybe, the lazy loading of the tags for the dialog plays a role...

Patch-set: 10
Label: Verified=0