Update patch set 1
Patch Set 1:

(5 comments)

> Should there be a special handling if "this version" is the working
 > tree version (compare current head without changes with arbitrary
 > commit)? Just wondering.

Perhaps, but I'd suggest if so this be done in a later commit. Should then probably also have the same behavior in the CommitFileDiffViewer's context menu in the history view.

 > * Is there a reason why "Open Working Tree Version" is available
 > only for the "this" version (both for ---a/+++b and @@ -u,v +x,y
 > @@)? I'd favour the same additional links for both. As a positive
 > side effect, you see where the "--- a/..." link goes (currently,
 > you don't because there is only one hyperlink).

Done.

 > * Is there a reason why "Open Working Tree Version" is listed
 > before "this version"? (proposed order of potential hyperlinks:
 > side by side compare, this/base version, working tree version)

Yes: I thought it would be the order of importance. Personally I use "Open Working Tree Version" most often, so maybe it should even be the top one. But I also quite frequently use the side-by-side link. In any case changing the order of menu items is always irritating to users who have to un-train old habits.

 > * I guess there is no way to force showing the hyperlink text if
 > there is only one hyperlink. For "diff -- git..." and "--- a/..."
 > there is only one hyperlink and you know where it leads only when
 > clicking.

Unfortunately this is hardcoded in MultipleHyperlinkPresenter.showHyperlinks() in JFace. Short of duplicating this class and installing our own instance that doesn't do "if (activeHyperlinks.length == 1) return;" I see no way to achieve that. Another approach is to open a tooltip ourselves, as in Mylyn's TaskHyperlinkPresenter, but that's a significant effort, too, and a lot of code for little benefit.

 > * When clicking on "diff --git...", the Compare editor opens. I
 > have the feeling that the editor title tooltip shows the commit ids
 > in the wrong order (or the two commits are passed in the wrong
 > order). In the non-swapped case, the base revision should be left
 > and this revision should be right, shouldn't it?

That compare side thing is confusing anyway. Traditionally Eclipse had a comparison between the workspace and some VCS version. The "local" version always was left. I think that's why EGit typically has the newer or local version left. (In "Compare with local history", local also is left and is always newer.) So "this" should be left, "base" should be right in the non-swapped case. Some titles are different, though. For the stand-alone DiffEditor, I chose the git standard base..tip, but the Compare editor says "Comparing tip with base", which is the other way around (but not wrong and in the same order as the panes if tip is left and base is right).

Patch-set: 1
Label: Verified=0