Update patch set 8
Patch Set 8:

In a separate change, I tried to investigate build failures in the TagActionTest. It seems that the bot started working when the tag tree was not fully populated. This is why more asserts were added when doing the selection.
My hope is that the test behaves more deterministic now and gives more helpful information if it fails.

Patch-set: 8