Update patch set 12
Patch Set 12:

The actual relevant changes here are in TagActionTest. All other changes were made to shorten the build time.

My finding is that possibly due to populating the tag tree in a asynchronous UI thread, the tree is not always fully populated when the swt bot is to make the tree item selection.

Patch-set: 12
Label: Verified=0