Update patch set 9
Patch Set 9:

> I noticed. I'll run this in a CentOS VM. It _is_ possible that
 > SWT/GTK has a bug. There were already problems earlier on, see the
 > comment in testConvertAnnotatedTagIntoLightWeight().

It is indeed the same problem. After the dialog opens, the tags are there:

  SomeTag
  SomeLightTag

but then selecting the second one via getTableItem("SomeLightTag") just doesn't work; it selects the first item. Whenever that happens, the tests go haywire.

So it's indeed the changed sort order that makes these tests now hit that problem.

I'll see if I can come up with a fix.

Patch-set: 9