Update patch set 9
Patch Set 9:

There's another problem in CreateTagDialog. In the annotated-to-lightweight case, the solution was to simply set the text field instead of selecting in the table. If I do the same in the lightweight-to-annotated case, the "Force overwrite" button never gets enabled. That also needs to be fixed.

Patch-set: 9
Label: Verified=0