WIP history view test stability 2
As waiting for jobs does not suffice, further attempts are made to
narrow the problem down.
In another unreliable UI-Test tree selection was one of the problems. So
the most relevant change is to do an unselect after checking the tree
items.

The test is duplicated, twice with doing the unselect, twice with.

Change-Id: I1b246552f7da25fee7cd1e16a4ce8b19ac9290c4
Signed-off-by: Alexander Nittka <alex@nittka.de>