WIP history view test stability 2
As waiting for jobs did not suffice, further attempts are made to narrow
the problem down.
In another unreliable UI-Test tree selection was one of the problems. So
the most relevant change is to do an unselect after checking the tree
items.

The main test does an unselect, the others have the
pre-first-stability-change state. In the best case, the main test never
fails and one of the others does.

Change-Id: I1b246552f7da25fee7cd1e16a4ce8b19ac9290c4
Signed-off-by: Alexander Nittka <alex@nittka.de>