Update patch set 1
Patch Set 1:

This is a follow up change to the compact outline tree feature. A review can already been done but a merge should wait for the decision on the compact tree.

I am unsure about the context for the keybinding (if it is OK/correct and what to do if it is not). At least for me, the keybinding worked as expected.

Patch-set: 1