Update patch set 1
Patch Set 1:

> This is a follow up change to the compact outline tree feature. A
 > review can already been done but a merge should wait for the
 > decision on the compact tree.
 > 
 > I am unsure about the context for the keybinding (if it is
 > OK/correct and what to do if it is not). At least for me, the
 > keybinding worked as expected.

Nice. Dialog appears in the middle of the screen; should appear in the middle of the editor area. Filter should support CamelCase and wildcards like in JDT Quick Outline.

Global window-wide context makes me nervous. Can't we have a context specific for the DiffEditor, or perhaps for the DiffViewer? Having it on the DiffViewer might perhaps make it easy to have this also in the history view...

Any reason why this has to be a static inner class? Can't we extract this into its own top-level class? Also, there's a TODO in that xtext code (relating to the user hitting return). Should be resolved.

Is there really no platform support for this? I'm a bit surprised to see JDT, Xtext, Mylyn, PDE, ..., and now EGit all re-implement this.

Patch-set: 1