Update patch set 1
Patch Set 1:

> ... should appear in the middle of the editor area.

I'll give it a try (or rather adapt the corresponding method from Xtext).

 > Filter should support CamelCase and wildcards like in JDT Quick Outline.

Yes, it is quick and dirty at the moment. I'll see if I can find code to copy.

 > Global window-wide context makes me nervous. Can't we have a
 > context specific for the DiffEditor, or perhaps for the DiffViewer?
 > Having it on the DiffViewer might perhaps make it easy to have this
 > also in the history view...

That's what I meant with I don't know how to fix the context part. I assume there is no existing DiffViewer context, I can simply use.

 > Any reason why this has to be a static inner class? Can't we
 > extract this into its own top-level class? 

My intention was to do as little refactoring as possible (because it is a follow-up and I have little experience with conflicts so far). It is pure luck that delegating to the existing OutlinePage worked so *well*. I'd like to clean up the (straight-forward) funcitonal issues and do the refactoring afterwards.

 > Also, there's a TODO in
 > that xtext code (relating to the user hitting return). Should be
 > resolved.

I added that TODO because I am not sure about the desired behaviour.
* A folder could already be selected in the tree.
* A diff could be selected in the tree.
* There could be multiple matching diffs.
* There could be a single matching diff.
For two out of these four, it is not immediately obvious to me what should happen.

Patch-set: 1
Label: Verified=0