Update patch set 1
Patch Set 1:

> That's what I meant with I don't know how to fix the context part.
 > I assume there is no existing DiffViewer context, I can simply use.

No there isn't. But shouldn't bee too difficult. Check the RepositoriesView. Is uses its own context and registers that with the IContextService obtained from the site. Additionally it uses a more local context that it activates/deactivates depending on what is selected.

Don't know off-hand where the DiffViewer would get the IContextService from; ideally it'd use the one from the site of the view or editor containing the viewer. If done for the Viewer, it might be necessary to activate/deactivate the context on focus events. If done for the DiffEditor, just registering with the IContextService obtained from the site should be fine.

Patch-set: 1
Label: Code-Review=0