Update patch set 1
Patch Set 1:

(2 comments)

General remark: I find the code of the run method hard to follow, in particular the purpose of the "if (commitToLoad != null) {"-part (lines 92ff) within the for-loop. It looks like it will run only in the first iteration if ever (commitToLoad is set to null).
If it is about making the operation cancellable, I still don't see the point because neither loadedCommits.fillTo(commitToLoad, maxCommits) (line 93) nor the loop for checking for the commit id (line 98) check the monitor. Only the cancellable batch loading (line 105) requires the loop!?

Patch-set: 1