Update patch set 1
Patch Set 1:

(2 comments)

> General remark: I find the code of the run method hard to follow,

So do I :-) It's ugly as hell. Basically, it tries to support incrementally loading a history, but as you pointed out, it has a number of flaws (such as finding a wanted commit only if it's in the first batch).

But this thing and the interactions with the GitHistoryPage is fragile as hell. Revamping this would be a lot of work; I wouldn't want to do that just to fix this bug.

Patch-set: 1