Update patch set 10
Patch Set 10:

(4 comments)

> Can you provide the toggling of the sort order in all the UI
 > elements where this is used ?

I would not expect to change the sort order for tags in the repositories view when using the Push tags dialog (or any of the other dialogs). And I also feel it is overkill to have a persisted toggle state for every dialog where tags are shown.

Personally, I'd make the sort order configurable only for the repositories view, but as stated in one of the file comment replies, I was asked to use the same sort order everywhere.

Patch-set: 10