Update patch set 1
Patch Set 1:

(2 comments)

> Revamping this would be a lot of work

I did not want you to do that. I really only wanted to say, I do not see at first glance what's happening.

Basically, I am fine with the change. My worry is that a fairly big problem is taken care of, leaving a tiny one, which may be even harder to track if it hits.

Patch-set: 1