Update patch set 2
Patch Set 2:

> Do I just add someone?

Yes. Or just let it sit and wait until someone comes along, but that's not really satisfying ;-) Or rebase it from time to time to make it perhaps pop up on whatever lists people are monitoring.

OTOH, having too many people in the reviewers list may be detrimental; it brings the risk that all think the others would review.

Normally Michael is pretty reponsive, but right now he seems to be gone for a while. Or just overloaded with other work. And BTW, yes, for a Windows double-check he was the right person to ask.

Patch-set: 2