Update patch set 2
Patch Set 2:

> Dialog appears in the middle of the screen;

Done.

 > Filter should support CamelCase and  wildcards

Still open.

 > Global window-wide context makes me nervous.

Done. I defined the context for the viewer. Currently, the action is only defined for the DiffEditor (and I don't really see the use case for the history view, as you have to select a file anyway), but this way the context activation should already be prepared when adding it elsewhere.

 > Any reason why this has to be a static inner class? Can't we
 > extract this into its own top-level class? 

Still open. And I am quite reluctant to do this refactoring. The inner classes are quite tightly connected (Content-/LabelProvider, Folder)... and it works so well with so few changes to the actual outline page code...

 > Also, there's a TODO in that xtext code

Done.

Patch-set: 2