RFC: [repo view] Tag filtering
Provide a new command that allows a user to filter tags by name in the
repositories view. The command is bound to the standard "Find" command
(Ctrl/Cmd-F) and is for now enabled only on the TagsNode. The command
expands the node and pops up a search text field providing near-
realtime filtering of the children of the "Tags" node.

Bug: 545240
Change-Id: I9393248990cdca951f2ff6ba71969f20ecd23f74
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>