Update patch set 1
Patch Set 1:

What do you think of this? It's not perfect yet; storing the filter pattern in the TagsNode means it's lost when a refresh occurs, which can also occur in response to background scanning for changes. But it shows the basic idea of an alternate solution for the problem in bug 545240: quickly find the interesting tags.

Patch-set: 1
Label: Verified=0