Update patch set 2
Patch Set 2:

(2 comments)

The position is much better now.

Normally the DiffEditorOutlinePage follows the selection in the DiffEditor. This appears to be broken after having used the quick outline to jump to a location: the outline doesn't highlight the new file, and doesn't update the selection either when the caret moves into another file inside the unified diff.

Patch-set: 2