Update patch set 3
Patch Set 3:

PS 3 simply attaches the text control to the tree instead of using a PopupDialog. On Mac, this makes the text field stick nicely with the table item and scroll with the tree. On gtk (of course!) it still doesn't, but it's no worse than the dialog at least.

Patch-set: 3
Label: Verified=0