Update patch set 4
Patch Set 4:

Just functional review: I like the feature very much!! 

The position of the search field is a bit strange depending on the width of the view. I'd like it to be right on or very close to the "Tags" label. Even if the view is not too wide, it is too much to the right (at least on Windows).

How about interpreting a find pattern -<number><d/w/m/y> as time restriction: -15d=tags of the last 15 days, -3m = last 3 months etc. Because the string matching alone would make it hard to filter tags version 4.9 onwards, this would be a *very concise* way to restrict the number of shown tags (see my original repo view filtering Bug 540003). I know this could turn into a filter language specification/support discussion (where to stop?). But I sure know I'd be sad if filtering on tag name only was possible.

Playing around with the egit tags I could also imagin restricting to *-r tags to be desirable (interpreting ^ and $ in the pattern).

Patch-set: 4