Update patch set 4
Patch Set 4:

> The position of the search field is a bit strange depending on the
 > width of the view.

Right, I tested only with the git repo view in the left stack, below the package explorer, and that isn't too wide in my setup. It should be possible to put this after the label text. 

 > How about interpreting a find pattern -<number><d/w/m/y> as time
 > restriction:

Then we're starting really to filter not by name but by commit date. But thinking more about this, I'm not sure it's needed. The recent tags should be shown in the git history view not too far down, shouldn't they?

But we could easily change this to handle ^ and $ (at the start or end of the pattern) or even general regular expressions. Or maybe even have a ToolItem next to the text field with a drop-down menu to select the mode (string, regexp, or date). One thing to consider, though, is that as soon as we implement some pattern matching language ("-15d", or regexps), we should also have some mechanism to display an error message if the pattern is invalid. Maybe again a tooltip, as in the MessagePopupTextCellEditor. And then the whole code definitely needs to be factored out into its own class.

BTW, I tried a CellEditor first, but that didn't work well. Maybe I should try again, though.

Patch-set: 4
Label: Verified=0