Update patch set 5
Patch Set 5:

PS 5: position of the text field improved, and (possibly anchored) wildcard matching.

 > If there was a
 > dropdown/code completion with something "last 10 day, last 4 weeks,
 > last 6 months, last 2 years" which then translates directly into
 > the text pattern

For that we would need to know what date format the tags used, and it would not find
recent tags that didn't follow that format.

I'd much rather we'd consider this again in another commit and then really filter by commit date. If it's really still needed now that wildcards are supported. (Still not the same as < or > some date, but for practical purposes it comes close enough I think.)

Patch-set: 5
Label: Verified=0