Update patch set 7
Patch Set 7:

(1 comment)

> For that we would need to know what date format the tags used, and
 > it would not find
 > recent tags that didn't follow that format.
 > 
 > I'd much rather we'd consider this again in another commit and then
 > really filter by commit date.

I am not opposed to dealing with that in a different change. And I am absolutely not opposed to filtering by commit date. In fact, I wasn't even thinking about the date of tagging (as we typically have a release build which commits a version update and tags that commit, so commit date and tag date are basically the same).

Patch-set: 7