Update patch set 7
Patch Set 7: Code-Review+1

(1 comment)

What do you think about persisting the filters? FilterChache already has all the needed information.
The tree expansion state is not persisted, but repository groups are and so are the other Pre-set filters, so it may be a matter of taste to which category these filters belong. I vote for persisiting. I see it as a more customizable form of the other filters - narrow down the view content to what I really need for working. And for projects with many tags this long list - with the wrong sorting ;-) - has always bothered be.
I would want to define a tag filter once and want it to stay defined until I change it again.

Thanks for following up on the sorting problem this way.

Patch-set: 7
Label: Code-Review=+1