Update patch set 4
Patch Set 4:

> If you are not generally opposed to the feature, layout and texts
 > can be adapted.

Of course not! It does make sense to give the user a way to specify the tag fetch option in the UI.

It's a fairly advanced feature, though; I doubt many people will change it from the default. In fact, there was a suggestion somewhere to allow skipping this page altogether... ah yes, it's bug 502558.

Patch-set: 4