Allow skipping clone destination page
The git clone wizard can be finished once the destination page has valid
values. With the old implementation those valid values are automatically
pre-filled when the destination page is set visible the first time. With
this change the clone destination values are pre-filled when the source
branch page becomes complete.

That way the user does not need to navigate to the (magically
pre-filled) destination page, but can instead finish from the branch
selection page.

Change-Id: I5abad43a8ee7fde501ffb82d2c4581cfdfdf1d62
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>