Update patch set 1
Patch Set 1:

Cool. Is it feasible to extend this change to bug 502558 skip the BranchSelectionPage as well? The cloning seems to support an "all-Branches" flag which could be set if the branch selection page was skipped. Currently this fails due to the first call of SourceBranchPage#checkPage setting the page to incomplete. Only if you go to the SourceBranchPage and back, the button is enabled.

Patch-set: 1