Update patch set 3
Patch Set 3:

(1 comment)

If there is a short cut to getting at the times, I'd be happy about a pointer. If there are no severe performance issues, I would not want to cache any results. After all, the previous code also went through all tags. Granted, now it will be slower due to the filter and I did not measure the impact for a large number of tags.
When printed out the timestamps of the tags, they were not ordered, so I hope the worst case (tags in chronological order) will not be the standard case.

Patch-set: 3
Label: Verified=0