Update patch set 1
Patch Set 1: Code-Review-1

This change does not handle the following user steps correctly

* open wizard with URL-1 (finish deactivated - OK)
* go to next page (branch selection)
* deselect one of many branches
* go back to Source Git Repository (finish activated - OK)
* enter repo URL-2 (finish still activated - not OK!?)
* press finish

There is an exception because the selected branches are from URL-1 but the cloning is done on URL-2. Even if you don't deactivate a branch and by default all branches are cloned, the clone destination directory and hence the repository name is wrong; the result of cloning not consistent.

One workaroud would be to ensure the cloneDestination-page to be incomplete on going back. However, my feeling is that it would be a better idea to have *all* the cloning relevant information kept in one object which would be shared among all pages and each page just modifies the parts it is responsible for. That way, all consistency checks and updates can be done in one place and the pages need not "update" each other. GitRepositoryInfo seems to be the obvious candidate.
When changing the repository-URL, branch selection and destination can be reset to default, default destination directory can be updated and checked for non-existence etc.

(I am not claiming that this refactoring is in the scope of this proposed change.)

Patch-set: 1
Reviewer: Gerrit User 236689 <236689@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1