Remove packed version of egit.doc from update site
egit.doc is the biggest contributor to the update site. The pack200
compressed version is almost as big as the not packed version. Therefore
do not even try to create a packed version of that bundle. This removes
5 MB from each update site that we publish.

This cannot be tested locally, since the signing profile cannot be
executed locally. Running with the default profile works fine locally.

Change-Id: I49908c706693d2bd01effb0716f596818effdc37
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>