GitCloneDropAdapter: correct CloneCommand invocation
Invoke the command through the IHandlerService instead of executing
it directly with a fake and incomplete ExecutionEvent.

This ensures that the command gets a proper event with all details
filled in and prevents an NPE that otherwise would occur during
command execution when the command tries to access the variable
for the current selection.

Bug: 562386
Change-Id: If080b1be111c13fc79e53aea31c07fe5c727b506
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>