Upgrade Tycho to 1.7.0
With this version the switch to JUnit 5 will be much easier, since Tycho
now supports the recent JUnit 5.6.0, and therefore the JUnit version is
finally the same in Maven/Tycho and in Eclipse.

At my workplace, this has been the first version of Tycho, where JUnit 5
tests work flawlessly both in Eclipse and in Maven.

Change-Id: Ie96786fd8da018578d6fd695d2cd11e55fde5a3e
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>