Update patch set 7
Patch Set 7:

Very cool, especially in combination with https://git.eclipse.org/r/#/c/162769/ . But Matthias is right: I have a repo with

  [remote "origin"]
    ...
    push = HEAD: refs/for/master

and a branch

  [branch "bar"]
    remote = origin
    merge = refs/heads/master
    rebase = true

where "Push HEAD..." gives me the "push branch" dialog irrespective of whether change-id is checked or not. (When it is checked, the button does have the "Gerrit repo" icon.)

It would be more consistent if it gave me the "Push to Gerrit" dialog since the config is set up that way. Perhaps the dialog logic from the CommitJob could be factored out and re-used in staging view directly?

Patch-set: 7
Label: Verified=0