Update patch set 7
Patch Set 7:

what if you have more remotes and want to push to another one than Gerrit?
The issue with the Gerrit wizard is that it doesn't allow to push to some non-Gerrit remote, and it can be annoying, while the generic Push Branch wizard can push to anything, including Gerrit.

I think before forcing for usage of the Gerrit push wizard, we should ensure there is a simple way to move from the Gerrit push to the regular push branch wizard.
Note that the wizard is always shown in case of Push HEAD..., so the remote is used to prepopulate the wizard, and using the Push Branch wizard wouldn't require any additional click or keystroke for the case you mention. So I'm even not convinced the Gerrit wizard is better in this case.

FWIW, I don't plan to work more on this issue soon. So anyone interested can merge and fix, or amend directly this commit in the meantime to provide such extra behavior.

Patch-set: 7
Label: Verified=0